+++ src/containers/sequence/list/list.h
@@ -47,7 +47,7 @@ struct _common_list {
 	void (*_reverse)(const common_list_t this);
 	errno_t(*_unique)(const common_list_t this,
 		errno_t(*comp)(const void* a, const void* b, bool* result),
-		int* removed);
+		size_t* removed);
 	errno_t(*_sort)(const common_list_t this,
 		errno_t(*comp)(const void* a, const void* b, bool* result));
 };
@@ -414,7 +414,7 @@ extern void common_list_reverse(const common_list_t this);
 * @return	On success, this function returns zero.  On error, an errno [...].
 */
 extern errno_t common_list_unique(const common_list_t this,
-	errno_t(*comp)(const void* a, const void* b, bool* result), int* removed);
+	errno_t(*comp)(const void* a, const void* b, bool* result), size_t* removed);
 
 /*
 * Sorts the elements in ascending order. The order of equal elements is
